New option linear_approximation for perfect foresight models.
This approach only requires one evaluation of the dynamic model (and its
jacobian) instead of T (the number of perdiods). Also (because the model
is linear) the equilibrium paths are obtained by inverting the jacobian
of the stacked equations (no need for a Newton algorithm).

Only available with stack_solve_algo==0 (which is the default algorithm
for solving perfect foresight models).

If possible, the option is triggered automatically if the model is
declared linear.

TODO:
 * Write a linear version of perfect_foresight_problem routine.
 * Evaluate the approxilation error (just need to evaluate the system of
 stacked non linear equations).