Closes #1057.
Use a static file with a list of routines to cleared before each run (because they have
persistent variables).This static file (the script
matlab/list_of_functions_to_be_cleared.m) is distributed with Dynare and
needs to be updated for each release (or each time we add/remove
persistent variables). To update the file just do:

>> clear_persistent_variables(DYNARE_MATLAB_PATH, true)

and add/commit the generated file.