Cosmetic change.
A "lag"" key with a negative value should be interpreted as a lead (and not a
lag). Possible solutions are (1) Change the sign of the value, (2) Add a new
key "lead" and consider only positive values, and (3) Change the key name for
"shift" with negative values for lags and positive values for leads.

The commit implements the third choice.