Revert "Allow th use of EXPECTATION keyword in variables tags."
This reverts commit fbc42b19301e918ccad0955b67ae814f4db9b70b, which is
incorrect for the currently used version of Bison (in stretch).
Besides, even if the patch seems to work with recent versions of
Bison, it introduces hard to predict and probably very unpleasant
properties (this patch allows to use the `expectation` word anywhere,
this would be a problem if the user choose to name a variable or
parameter as `expectation`).