Skip constant equations replacement only when there is an “mcp” tag
Since bef537d40accd597c19f112471974b726c9c9890, constant equations were not
simplified as soon as they had a tag attached.

But this is too wide a restriction. In particular, this breaks the trend
component models which have a target that is set to a constant.

So we now only skip the replacement in the case where there is an “mcp” tag.

Ref. dynare#1697