No longer replace all auxiliary variables by their definition in the static model
This is effectively a revert of commits 1b4f68f93433cf1fd826789f32a567c34f4c1e7d,
32fb90d5f3c729f0744ad97ffc7c51e31cb3731b and f6f4ea70fbce9fd584c72c15a2932fa02a854d0b.

This transformation had been introduced in order to fix the computation of the
Ramsey steady state in the case where Lagrange multipliers appeared with a lead
or lag ⩾ 2 (and where thus part of the definition of an auxiliary variable).

But this transformation had introduced bugs in the handling of external
functions which were difficult to tackle.

Moreover, it seems preferable to keep the strict correspondence between the
dynamic and static model, in order to make reasoning about the preprocessor
internals easier (in particular, for this reason this transformation was not
implemented in ModFile::transformPass() but in ModFile::computingPass(), which
was a bit confusing).

A better solution for the Ramsey steady state issue will is implemented in the
descendent of the present commit.

Ref. dynare#633, dynare#1119, dynare#1133