🐛 Flex scanner: fix column location counter in the presence of dates
Dates handling inside Dynare statements was done in the Flex scanner using a
hack that would introduce some extra characters in the character stream (in
order to return a call to the “dates('…')” constructor to Bison). But this
would break the column location counter, because the latter would count
characters that do not exist in the user file.

This commit fixes it by implementing a more natural handling of dates, adding
the “dates('…')” constructor directly at the level of the semantic value.