Refactor getJacobianCol() and getJacobianColsNbr() methods
This follows the removal of the legacy representation. These methods now only
work according to the new (sparse) representation. By the way, they are moved
to the ModelTree class and marked as [[nodiscard]].

A new getLegacyJacobianCol() method is created for the two outputs that still
need it: bytecode and M_.lead_lag_incidence.

Ref. dynare#1859