Flex lexer: use a more explicit token name for an identifier
So far, the token name for an identifier was NAME.

This was confusing because this could easily be mistaken for the “name”
literal. And actually, there was a bug in “occbin_constraints”, “init2shocks”
and “shock_groups”, which would accept any identifier in place of the “name”
literal that allow in their options/body.

We now use the IDENTIFIER token name for identifiers, and NAME for the “name”
literal.

As a consequence, the EXT_FUNC_NAME token is no longer needed (it was used for
the “name” literal in the context of the “external_function” statement).

Also, for consistency, rename the EXT_FUNC_NARGS token to NARGS.